authentification OK (correctif à apporter => fermeture des alertes bootstrap)

+++ src/server.js
@@ -9,7 +9,7 @@ const AuthController = require('./Controllers/AuthController');
 const express = require("express");
 const swaggerJSDoc = require('swagger-jsdoc');
 const swaggerUi = require('swagger-ui-express');
-const bodyParser = require('body-parser');
+
 
 
 // Initialization variables
@@ -37,14 +37,14 @@ module.exports = swaggerSpec;
 // Initialize swagger-jsdoc -> returns validated swagger spec in json format
 app.use('/api-doc', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 app.use(session({secret:'secretkey'}));
-app.use((req,res,next) => {
-  console.log(req.url);
-  if(req.url !== '/login.html' && req.url !== '/' && req.url !== '/css/style.css' && req.url.includes('api'))
-    if(!req.session.token)
-      res.status(401).send();
 
-  next();
-});
+const isAuth = (req) => {
+  const tokenSession = req.session.token;
+  const tokenAuth = req.get('Authorization');
+  if(!tokenSession && !tokenAuth)
+    return false;
+  return tokenSession === tokenAuth;
+};
 app.use(express.static('./public'));
 app.use(express.json()); // for parsing application/json)
 app.listen(server_port);
@@ -52,6 +52,15 @@ app.listen(server_port);
 app.get('/', async (req,res) => {
   res.redirect('/login.html');
 });
+
+app.get('/promocodes', (req,res) => {
+  if(!isAuth(req)) {
+    res.redirect('/login.html');
+    return;
+  }
+  res.redirect('/promocode.html');
+});
+
 /**
  * @swagger
  * 
@@ -81,12 +90,26 @@ app.get(api + "/coupon/:id", async (req, res) => {
   res.append("Content-Type", "application/json");
   dbclient.getCodePromoByQrCodeId(req.params.id)
     .then((result) => {
-      res.send({ codepromo: result.code });
+      res.send(result);
     }).catch((erreur) => {
       res.status(404).send({ erreur: "Not Found" });
     });
 });
 
+
+app.get(api + "/coupons", async (req, res) => {
+  if(!isAuth(req)) {
+    res.redirect(401,'/unauthorization.html');
+    return;
+  }
+  res.append("Content-Type", "application/json");
+  dbclient.getCodesPromos()
+      .then((result) => {
+        res.status(200).send(result);
+      }).catch((erreur) => {
+        res.status(404).send({ erreur: "Not Found" });
+      });
+});
 /**
 * @swagger
 * 
@@ -119,12 +142,14 @@ app.get(api + "/coupon/:id", async (req, res) => {
 *          description: Récupération OK
 */
 app.post(api + "/coupon", async (req, res) => {
-  console.log(req.body);
-  const body = JSON.parse(req.body);
-  console.log(body);
-  dbclient.postCodePromo(body)
+  console.table(req.body);
+  if(!isAuth(req)) {
+    res.redirect(401,'/unauthorization.html');
+    return;
+  }
+  dbclient.postCodePromo(req.body)
     .then((result) => {
-      res.send({ message: result });
+      res.status(200).send(result);
     }).catch((erreur) => {
       res.status(500).send({ erreur: erreur });
     });
@@ -133,12 +158,17 @@ app.post(api + "/coupon", async (req, res) => {
 app.post("/login", async (req, res) => {
   res.append("Content-Type", "application/json");
   req.session.token = null;
+  console.table(req.body);
   const token = await authController.login(req.body.login, req.body.password);
   console.log({token});
-  if(!token) res.redirect('/');
+  if(!token) {
+    res.redirect(404,'/');
+  }
   else{
     req.session.token = token;
-    res.redirect('/promocode.html');
+    res.status(200).send(JSON.stringify({
+      token: token
+    }));
   }
 });
 
@@ -147,17 +177,12 @@ app.post(api + "/logout", async (req, res) => {
   req.session.token = null;
 });
 
-app.get(api + "/coupons", async (req, res) => {
-  res.append("Content-Type", "application/json");
-  dbclient.getCodesPromos()
-      .then((result) => {
-        res.status(200).send(result);
-      }).catch((erreur) => {
-          res.status(404).send({ erreur: "Not Found" });
-  });
-});
 
 app.delete(api + "/coupon/:id", async (req,res) => {
+  if(!isAuth(req)) {
+    res.redirect(401,'/unauthorization.html');
+    return;
+  }
   res.append("Content-Type", "application/json");
   dbclient.deleteCodePromo(req.params.id)
       .then((result) => {
